requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-dotenv==1.1.1
GitPython==3.1.34
haystack-ai==2.16.0
pgvector-haystack
ragas-haystack
voyage-embedders-haystack
sentence-transformers>=5.0.0,<6.0.0
transformers>=4.54.1,<5.0.0
psycopg[binary]
asyncpg
passlib
pyjwt
pydantic-settings>=0.2.2
pydantic[email]
tree_sitter==0.20.4
tree_sitter_languages==1.6.1
jose
aiofiles
datasets
langchain-community
langchain-groq 
mkdocs>=1.6
mkdocs-material>=9.0
mkdocstrings[python]>=0.24
mkdocs-pdf-export-plugin>=0.5
mkdocs-gen-files
mkdocs-literate-nav